requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==5.0.6
django-heroku==0.3.1
django-import-export==4.0.3
gunicorn==22.0.0
packaging==24.0
psycopg2==2.9.9
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.0
tzdata==2024.1
whitenoise==6.6.0
